[#45] Set up Travis CI (#64)
This is to allow all future PRs to be tested on the travis CI
service for the convenience of the project maintainers.
This ensures that PRs does not  break any of the existing
features.

The build file will instruct travis to write a log message before
running any test. This is to give developers and maintainers
looking at the log an ease of mind that the tests are
indeed being run. This is because the log will look very empty
and suspicious even if the build passed if this feature is not
implemented.