UserGuide: Remove section "Changing Save Location"
The UserGuide claims that the save location of the addressbook data can
be changed via command line parameters.

That feature is not fully implemented. As such, addressbook data is
always saved in addressbook.txt and cannot be changed.

Let's remove the section "Changing Save Location". We can restore this
section after the feature is fully implemented.