DeveloperGuide: Change setup to use default test configuration
The working directory needs to be changed for the tests to run properly.
Therefore, the developer guide instructs the developer to modify the
generated test configuration.

If the IDE generates another new test configuration, the settings needs
to be changed again, otherwise the tests will fail again.

Let's change the instructions to ask the developers to modify the
default test configuration instead of the generated one.