DeveloperGuide: update instructions to import project as gradle project
The instructions for users for setting up of the project in Intellij
is to import the project using the root folder.

In order to lessen the learning curve between AB3 and AB4, as well
as to import necessary 3rd party dependencies more easily, we should
update the instructions to import the project as a gradle project
instead.

Let's update the instructions in DeveloperGuide to import the project
into Intellij as a gradle project.