AddressBook: remove master tag list
We keep track of a master tag list which holds all the tags in
AddressBook.

As the master tag list was deemed unnecessary and may be too complex
for new developers to grasp, as discussed in
se-edu/addressbook-level4#753 and se-edu/addressbook-level4#794,
the master tag list has already been removed in addressbook-level4
in se-edu/addressbook-level4#825. As such, the master tag list
should also be removed here for consistency.

Let's remove the master tag list in AddressBook.

Similarly, it is now unnecessary for AdaptedAddressBook to store
a master tag list as well.

Let's remove the master tag list in AdaptedAddressBook.