docs/mainClassDiagram: remove UniqueTagList
AddressBook previously had a UniqueTagList linked to it as the
master tag list, which was reflected in the mainClassDiagram
architecture diagram in the DeveloperGuide.

Since AddressBook does not use a master tag list anymore, and
UniqueTagList has been removed, the architecture diagram is no
longer consistent with the program.

Let's update the mainClassDiagram architecture diagram.