Upgrade project to run on Java 9
The project currently runs on Java 8.

Since AddressBookL4 is already running on Java 9, this project should
be upgraded to run on Java 9 as well for consistency.

Let's upgrade the project to run on Java 9 and update the
documentation accordingly.

Travis currently builds the project using JDK 8. Since we are now
running on Java 9, let's update it to build the project using JDK 9.

Java 9 introduces a module system, causing internal packages such as
javax.* and com.sun.* packages to be inaccessible[1]. As our project
uses these packages, it is no longer compilable. To resolve this,
let's update build.gradle to import these modules.

UtilsTest uses the Integer(int) constructor for its equality test,
which is deprecated from Java 9[1]. Since use of deprecated APIs is
considered bad practice, let's replace this with Integer.valueOf(int).

[1]: http://openjdk.java.net/jeps/260