Upgrade project to run on Java 9
The project currently runs on Java 8.

Since AddressBookL4 is already running on Java 9, this project should
be upgraded to run on Java 9 as well for consistency.

Let's upgrade the project to run on Java 9 and update the
documentation accordingly.

Travis currently builds the project using JDK 8. Since we are now
running on Java 9, let's update it to build the project using JDK 9.

Java 9 introduces a module system, causing internal packages such as
java.xml.bind to be inaccessible to unnamed modules by default[1]. As
our project is still an unnamed module which requires these packages,
it is no longer compilable. To resolve this, let's update build.gradle
to import these packages.

UtilsTest uses the Integer(int) constructor for its equality test,
which is deprecated from Java 9[2]. Since use of deprecated APIs is
considered bad practice, let's replace this with Integer.valueOf(int).

[1]: http://openjdk.java.net/jeps/261
[2]: https://docs.oracle.com/javase/9/docs/api/java/lang/Integer.html#Integer-int-