DeveloperGuide: add "Troubleshooting" sub-section
Due to an ongoing issue[1] with some of the newer versions of
IntelliJ (Ultimate Edition), the IDE may report code errors for
JAXB APIs, although the code can be build and run successfully
since we have already imported the dependencies through gradle.

This may cause confusion amongst users.

Let's add a "Troubleshooting Setting up" sub-section within the
"Setting up" section to address this issue, and how to resolve it.

[1]: https://youtrack.jetbrains.com/oauth?state=%2Fissue%2FIDEA-189060